Fixed set filter areas. Updated translations.

+++ helper.h
@@ -5,14 +5,14 @@
 #include <QDebug>
 
 namespace Helper {
-    inline float convertCoordinateInDec(QString coordStr)
+    inline double convertCoordinateInDec(QString coordStr)
     {
         QRegExp regExp("(\\D?)(1\\d{2}|0\\d{2}|\\d{2})\\s?(\\d{2})\\s?(\\d{2}(\\,|\\.\\d{1,2})?)");
-        float dec = 0;
+        double dec = 0;
 
 
         if (regExp.indexIn(coordStr) != -1) {
-            dec = regExp.cap(2).toInt() + regExp.cap(3).toFloat() / 60 + regExp.cap(4).replace(",", ".").toFloat() / 3600;
+            dec = regExp.cap(2).toInt() + regExp.cap(3).toDouble() / 60 + regExp.cap(4).replace(",", ".").toDouble() / 3600;
             if (regExp.cap(1).contains(QRegExp("[юз]")))
                 dec *= -1;
         }
